services/movies-service/requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
requests==2.31.0
PyJWT[crypto]==2.8.0
redis==5.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pika==1.3.2